detour penalty
