+++ tests/back/integration/test_008b_group_day_trips_congestion_changes_costs.py
@@ -49,7 +49,6 @@ def test_008b_group_day_trips_congestion_changes_costs(test_data):
         parameters=Parameters(
             n_iterations=1,
             n_iter_per_cost_update=0,
-            alpha=0.01,
             dest_prob_cutoff=0.9,
             k_mode_sequences=6,
             cost_uncertainty_sd=1.0,
@@ -86,7 +85,6 @@ def test_008b_group_day_trips_congestion_changes_costs(test_data):
         parameters=Parameters(
             n_iterations=2,
             n_iter_per_cost_update=1,
-            alpha=0.01,
             dest_prob_cutoff=0.9,
             k_mode_sequences=6,
             cost_uncertainty_sd=1.0,
