+++ tests/back/integration/test_008c_group_day_trips_parameter_profiles_change_iteration_2.py
@@ -48,7 +48,6 @@ def test_008c_group_day_trips_parameter_profiles_change_iteration_2(test_data):
         parameters=Parameters(
             n_iterations=2,
             n_iter_per_cost_update=0,
-            alpha=0.01,
             dest_prob_cutoff=0.9,
             k_mode_sequences=6,
             cost_uncertainty_sd=1.0,
@@ -77,7 +76,6 @@ def test_008c_group_day_trips_parameter_profiles_change_iteration_2(test_data):
         parameters=Parameters(
             n_iterations=2,
             n_iter_per_cost_update=0,
-            alpha=0.01,
             dest_prob_cutoff=0.9,
             k_mode_sequences=6,
             cost_uncertainty_sd=1.0,
