+++ tests/back/integration/test_010_group_day_trips_results_are_reproducible.py
@@ -33,7 +33,6 @@ def test_010_group_day_trips_results_are_reproducible(test_data):
         parameters=Parameters(
             n_iterations=1,
             n_iter_per_cost_update=0,
-            alpha=0.01,
             dest_prob_cutoff=0.9,
             k_mode_sequences=3,
             cost_uncertainty_sd=1.0,
@@ -55,7 +54,6 @@ def test_010_group_day_trips_results_are_reproducible(test_data):
         parameters=Parameters(
             n_iterations=1,
             n_iter_per_cost_update=0,
-            alpha=0.01,
             dest_prob_cutoff=0.9,
             k_mode_sequences=3,
             cost_uncertainty_sd=1.0,
