+++ tests/back/integration/test_011_population_trips_behavior_change_phases_can_be_computed.py
@@ -36,7 +36,6 @@ def test_011_population_trips_behavior_change_phases_can_be_computed(test_data):
         parameters=Parameters(
             n_iterations=3,
             n_iter_per_cost_update=0,
-            alpha=0.01,
             dest_prob_cutoff=0.9,
             k_mode_sequences=3,
             cost_uncertainty_sd=1.0,
