+++ tests/back/unit/domain/group_day_trips/test_003_behavior_change_scopes.py
@@ -1,12 +1,10 @@
 import math
 from pathlib import Path
-from types import SimpleNamespace
 
 import polars as pl
 import pytest
 
 from mobility.trips.group_day_trips import BehaviorChangePhase, BehaviorChangeScope, Parameters
-from mobility.trips.group_day_trips.plans.destination_sequences import DestinationSequences
 from mobility.trips.group_day_trips.plans.candidate_plan_steps import CandidatePlanStepsAsset
 from mobility.trips.group_day_trips.plans.plan_ids import add_plan_id
 from mobility.trips.group_day_trips.plans.plan_updater import PlanUpdater
@@ -114,209 +112,6 @@ def test_parameters_resolve_behavior_change_scope_from_active_phases():
     assert Parameters().get_behavior_change_scope(1) == BehaviorChangeScope.FULL_REPLANNING
     assert parameters.get_behavior_change_scope(3) == BehaviorChangeScope.MODE_REPLANNING
     assert parameters.get_behavior_change_scope(6) == BehaviorChangeScope.DESTINATION_REPLANNING
-
-
-def test_sample_active_destination_sequences_keeps_only_active_activity_sequences(tmp_path):
-    class _StubAsset:
-        def __init__(self, df):
-            self._df = df
-
-        def get_cached_asset(self):
-            return self._df
-
-    destination_sequences = DestinationSequences(
-        run_key="run",
-        is_weekday=True,
-        iteration=3,
-        base_folder=_make_local_tmp_path(tmp_path, "active_destination_sequences"),
-        current_plans=pl.DataFrame(
-            {
-                "demand_group_id": [1],
-                "activity_seq_id": [10],
-                "time_seq_id": [1],
-                "dest_seq_id": [100],
-                "mode_seq_id": [1000],
-            },
-            schema={
-                "demand_group_id": pl.UInt32,
-                "activity_seq_id": pl.UInt32,
-                "time_seq_id": pl.UInt32,
-                "dest_seq_id": pl.UInt32,
-                "mode_seq_id": pl.UInt32,
-            },
-        ),
-        activity_sequences=_StubAsset(
-            pl.DataFrame(
-                {
-                    "demand_group_id": [1],
-                    "activity_seq_id": [10],
-                    "time_seq_id": [0],
-                    "seq_step_index": [0],
-                    "activity": ["work"],
-                },
-                schema={
-                    "demand_group_id": pl.UInt32,
-                    "activity_seq_id": pl.UInt32,
-                    "time_seq_id": pl.UInt32,
-                    "seq_step_index": pl.UInt32,
-                    "activity": pl.Utf8,
-                },
-            )
-        ),
-        activities=[],
-        transport_zones=None,
-        destination_saturation=pl.DataFrame(),
-        demand_groups=pl.DataFrame(),
-        costs=pl.DataFrame(),
-        parameters=Parameters(),
-        seed=123,
-    )
-
-    seen = {}
-
-    def fake_run(activities, transport_zones, destination_saturation, chains, demand_groups, costs, parameters, seed):
-        seen["chains"] = chains
-        return pl.DataFrame(
-            {
-                "demand_group_id": [1],
-                "activity_seq_id": [10],
-                "dest_seq_id": [100],
-                "seq_step_index": [0],
-                "from": [1],
-                "to": [2],
-                "departure_time": [8.0],
-                "arrival_time": [9.0],
-                "next_departure_time": [17.0],
-                "iteration": [3],
-            },
-            schema={
-                "demand_group_id": pl.UInt32,
-                "activity_seq_id": pl.UInt32,
-                "dest_seq_id": pl.UInt32,
-                "seq_step_index": pl.UInt32,
-                "from": pl.Int32,
-                "to": pl.Int32,
-                "departure_time": pl.Float64,
-                "arrival_time": pl.Float64,
-                "next_departure_time": pl.Float64,
-                "iteration": pl.UInt32,
-            },
-        )
-
-    destination_sequences.run = fake_run
-    destination_sequences._sample_active_destination_sequences()
-
-    assert seen["chains"].select("activity_seq_id").to_series().to_list() == [10]
-
-
-def test_destination_probability_inputs_use_cost_and_sink_even_without_destination_utilities(tmp_path):
-    destination_sequences = DestinationSequences(
-        run_key="run",
-        is_weekday=True,
-        iteration=1,
-        base_folder=_make_local_tmp_path(tmp_path, "destination_utilities_default_zero"),
-        activities=[],
-        transport_zones=None,
-        destination_saturation=pl.DataFrame(),
-        demand_groups=pl.DataFrame(),
-        costs=pl.DataFrame(),
-        parameters=Parameters(),
-        seed=123,
-        resolved_activity_parameters={},
-        current_plans=pl.DataFrame(),
-    )
-
-    class _TransportCosts:
-        def __init__(self):
-            self.modes = [SimpleNamespace(inputs={"parameters": SimpleNamespace(name="car")})]
-
-        def get_costs_by_od_and_mode(self, columns, detail_distances=False):
-            return pl.DataFrame(
-                {
-                    "from": [1],
-                    "to": [10],
-                    "mode": ["car"],
-                    "cost": [3.0],
-                    "distance": [10.0],
-                    "time": [1.0],
-                }
-            )
-
-    opportunities = pl.DataFrame(
-        {
-            "to": [10],
-            "activity": ["work"],
-            "opportunity_capacity": [100.0],
-            "k_saturation_utility": [1.0],
-        }
-    ).with_columns(activity=pl.col("activity").cast(pl.Enum(["work"])))
-
-    costs_by_bin, cost_bin_to_destination = destination_sequences._get_destination_probability_inputs(
-        opportunities=opportunities,
-        costs=_TransportCosts().get_costs_by_od_and_mode(["cost", "distance", "time"]),
-        cost_uncertainty_sd=1.0,
-    )
-
-    assert costs_by_bin.collect().height > 0
-    assert cost_bin_to_destination.collect().height > 0
-
-
-def test_reuse_current_destination_sequences_reuses_current_plan_steps(tmp_path):
-    destination_sequences = DestinationSequences(
-        run_key="run",
-        is_weekday=True,
-        iteration=4,
-        base_folder=_make_local_tmp_path(tmp_path, "reuse_current_destination_sequences"),
-        current_plans=pl.DataFrame(
-            {
-                "demand_group_id": [1],
-                "activity_seq_id": [10],
-                "time_seq_id": [1],
-                "dest_seq_id": [100],
-                "mode_seq_id": [1000],
-            },
-            schema={
-                "demand_group_id": pl.UInt32,
-                "activity_seq_id": pl.UInt32,
-                "time_seq_id": pl.UInt32,
-                "dest_seq_id": pl.UInt32,
-                "mode_seq_id": pl.UInt32,
-            },
-        ),
-        current_plan_steps=pl.DataFrame(
-            {
-                "demand_group_id": [1, 1],
-                "activity_seq_id": [10, 10],
-                "time_seq_id": [1, 1],
-                "dest_seq_id": [100, 100],
-                "mode_seq_id": [1000, 1000],
-                "seq_step_index": [0, 1],
-                "from": [21, 22],
-                "to": [22, 23],
-                "departure_time": [8.0, 9.0],
-                "arrival_time": [8.5, 9.5],
-                "next_departure_time": [9.0, 17.0],
-            },
-            schema={
-                "demand_group_id": pl.UInt32,
-                "activity_seq_id": pl.UInt32,
-                "time_seq_id": pl.UInt32,
-                "dest_seq_id": pl.UInt32,
-                "mode_seq_id": pl.UInt32,
-                "seq_step_index": pl.UInt32,
-                "from": pl.Int32,
-                "to": pl.Int32,
-                "departure_time": pl.Float64,
-                "arrival_time": pl.Float64,
-                "next_departure_time": pl.Float64,
-            },
-        ),
-    )
-
-    result = destination_sequences._reuse_current_destination_sequences()
-
-    assert result["iteration"].unique().to_list() == [4]
-    assert result["seq_step_index"].sort().to_list() == [0, 1]
 
 
 def test_get_transition_probabilities_blocks_stay_home_in_mode_replanning(tmp_path):
